modify plugin to not pollute global states
This change  removes  languageVm as a singleton and instead favors a vue-router style injection approach.  This allows for the plugin to hopefully be able to be fully utilized for Server Side Rendering while still supporting all of the directives, and most of the original functionality.

relocate install function into new file called 'install.js'

refactor index.js to contain a class called VueGettext.

The majority of these refactors are to fall in line with the vue plugin style approach as demonstrated by vue-router, and vue-i18n.

modify root Vue instance to expect a gettext object that is an instanciated VueGettext object.

remove filter section from development demonstration.

Filters are unable to reference the vue node, and don't seem to have a vue binding directly.  A separate VueGettext could be re-instaciated if the filter is absolutely necessary, however it doesn't seem to fit with the spirit of what filter is ultimately trying to do (from a Vue perspective)

remove the ability to pass in a language mixin

This choice is more of a lack of understanding on my part.  I don't really understand why it is necessary, when a user could just create a mixin of their own using the Vue standard syntax.  However, it's likely also due to not beign able to have a nested object anymore.

rename $language.current to $currentLanguage

rename $language.available to $availableLanguages

These changes are motivated by the bindings dictated by Object.defineProperty.  When attempting to nest an object, it requires that the property be defined on the prototype.  It's possible to bind the Vue object instance, I suppose.  However, giving direct access to that also seems unecessary.  Instead providing access to the $currentLanguage and $availableLanguages seems like a much better choice.  As such the $language object is no longer available.

modify test cases to match the new instanciation and language invocation behavior

remove mixin test case